+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and to the [CHANGELOG recommendations](http://keepachangelog.com/).
 - Add Carrier bundle hatchback field to booking admin
 - Add listing search engine on user profile page
 - Add error message on listing calendar edition
+- Set booking status to STATUS_PAYMENT_REFUSED when booking can not be validated
 
 ### Fixed
 - Fix js escaping on result